Adding in space and time to model
This new commit adds in a branch of the model that will predict over day of year (for now) and space. The new filesystem is divided by these branches, and the filepaths that pull from the Data folder will need to be updated accordingly. Also, the aggregation functions in ...Temp Comparison.py have been moved to a functions folder, which the space_and_time script now imports. In the future, the space script will also import from here.