Parameter Optimization
Parameter Optimization to determine kernel and noise for training the time model. Results are added in the Results/ folder. Training is computationally expensive (likely as data is left intentionally unstandardized to allow for differences in how kernels treat spatial and time variables), but the results conclusively point to an alpha level of about a quarter of a degree and the benefit of using a rational quadratic kernel in addition the anisotropic RBF.