Further Model Development
Seperated the scripts that fuse data with coastal features from the Geostatistics Workbooks (in both the space and time model). Also created a geostatics prediction dashboard for the time model, but found that using any combination of kernels while training space and time variables together doesn't work. Either the model only works when the spatial variables are on the extreme of kernel means (demeaning destroys spatial prediction) or the model reduces to predicting one time series for the entire sound. The next thing to try will be sequential learning.