Made Time Dependent Model Work
Previous commit message was wrong, the standardization didn't destroy predictability, it was just set up wrong. Now fixed for the space version of the model (Geostatistics Workbook 3) and also for the new time version of the model (Geostatistics_Workbook_Time_3.1). Presumably results will now align with the parameter optimization results (soon to be uploaded).